refactor(aigc-video-gen): 拆分三供应商脚本 + MiniMax H3 多模态参考支持
将原 scripts/gen.py 拆分为三个独立供应商脚本，sh 层根据 --platform 或
env 自动 dispatch：

- scripts/gen_minimax.py — MiniMax Hailuo 视频生成 + music 子命令
- scripts/gen_volc.py — 火山引擎 Seedance 视频生成
- scripts/gen_dashscope.py — 阿里云百炼 HappyHorse / Wan2.7 视频生成
- skills/_shared/aigc_common.py — 共享逻辑（HTTP/download/ensure_safe_output/
  extract_last_frame/append_decision/generate/resolve_prev_segment）
- aigc-video-gen.sh — 重写为 dispatch wrapper（剔除 --platform 参数后转发）

MiniMax H3 多模态参考增强（对齐官方 video-generation-v2-h3-context-ir 文档）：

- 新增 --ref-audio 参数（reference_audio role）
- mm_video_build_content 支持 reference_image/reference_video/reference_audio 组合
- cmd_video 入口校验官方约束：
  - 图生视频（first_frame/last_frame）与多模态参考（reference_*）互斥
  - 多模态参考不可仅输入音频，须至少 1 个 reference_video 或 reference_image

SKILL.md 同步更新：三脚本结构说明、--ref-audio 参数行、H3 多模态参考约束章节。

验证：
- py_compile 四个 Python 文件全通过
- sh dispatch：--platform minimax 正确剔除并转发；env 自动判 MINIMAX_API_KEY →
  minimax；全无 env → exit 2 + 正确 hint
- H3 约束校验：仅 --ref-audio → die（须至少 1 个参考视频/图片）；
  --ref-audio + --ref-image → 通过约束进入 submit（401 假 key 预期）；
  图生视频 + 多模态参考 → die（互斥）
